extract the generation of the properties file into a separate class
